Code drop
Contains the MQTT broker with device twin support, a REST API
for the twin with device authorization, and a generic backend
as a service which is configured with a simple JSON file.

An example 'fleet' demonstrates usage. It comes with a node.js
app to add some data, and a python2 based device emulator.