fix created_at index
from the postgres doc:
The sort expression(s) can be any expression that would be valid in the query's select list. An example is:
SELECT a, b FROM table1 ORDER BY a + b, c;
When more than one expression is specified, the later values are used to sort rows that are equal according to the earlier values.

i.e. we don't need the index for the uuid. The alternative would have
been to create yet another index for created_at only, but I would like
to avoid too many indices (we already have plenty)