fix client for create
POST creates a new resource when the primary is not known.
PUT updates a resource with the primary id known.
But PUT can also create resources with the primary id given

The client therefore must select POST or PUT for create, depending
on whether it has a primary id or not.