Make concurrency model solid
A postgres database by default handles 100 concurrent connections.
In reality it can only work on one query per core concurrently.
Given that our processing is mostly database-bound, and the database
currently is running on a small instance,  it does not
make sense to have more than 10 concurrent processes.

Therefore we do the following setup:
1 the fitness lambda supports up to 10 concurrent http requests
2 the fitness-sns lambda supports up to 2 concurrent instances, each
  processing with 5 concurrent go-routines.

when we move to a larger db, we can increase the number of lambda
instances, and we can set a higher PipelineConcurrency level in the
backend (currently 5).