fix external activity ID
we cannot rely on external activity id from strava, hence
we use the strava itself as external id.

made activity import more resiliant, if an activity cannot
be imported, continue with the remaining ones

update beats formula to not use %HR but %HR reserve
as a factor (assuming resting hr is 60 and max is 180).
This makes sure that long walks do not give excessive
beat numbers.