support scheduled events
we now have RaiseEvent and ScheduleEvent, where ScheduleEvent
takes an additional scheduleAt timestamp.

As an example we create a morning message for every athlete for
the next day when they interact with the coach.

The change required some refactoring of context handling and events.
The code now looks more reliable and easier to read.

There is also a new kurbisio/event/send-message-to-athlete.
Put a models.MessageProperties into the body and do a PUT with
?resource=athlete&resource_id={your_athlete_id}