job handling fixes plus testing
this creates a dummy and dummy_shadow resources, and a
handler in the fitness service which copies dummy over
to dummy_shadow based on the notifications.

A test client service/fitness/tools/testing/main.go creates,
updates and deletes dummy objects and validates that they
get copied to dummy_shadow. It then prints out performance
stats.

Creating the test required low-level fixes in the
notifications system.