add default properties to configuration
Any Singleton or Collection resource can have an additional property "default", which defines default properties for
all instances. Default properties are automatically added whenever objects are created or updated in the database.
In addition, they are also added when older versions of objects are read from the database. Default properties
are especially useful in combination with schema validation, as they make it possible to add new required properties
without having to migrate all existing objects in the database.