fix singleton collection usecase
singletons no longer have their own ID but use their parent's ID
as main identifier. The collection handling was always skipping the
ID  to get to a collection, but for a singleton we must not do this.
The fix introduces a variable ownerIndex which stores the index
of the owner id, which is 0 in the singleton case and 1 in the
case of an ordinary collection.