Fix permission coorruption when using relation
When a a/b relation is created, a virtual a/b resource is also created.
The permission associated to this virtual resource is copied from the
original a and b permits. Yet since permissions are stored as pointers
and no deep copy is performed, this corrupted the original permission of a and b