fix for the jwt middelware
sometimes the jwt may simply not contain an email claim. This can
happen with apple sign in on firebase, when the email adress is
completely hidden. In this case we must still create a valid and
unique identity, by replacing the email claim with the unique
id that exists in the jwt.