introduce new relation resources
Relations are a special type of collection, which creates a relation between two resources.
For example, a relation "device_ownership" creates a relation between "user" on the left and "device" on the right.

The generated routes are as follows:

	/device_ownerships GET,PUT,PATCH,DELETE with optional query parameters ?left={user_id} and ?right={device_id}
	/device_ownerships/{users_id}:{device_id} GET,PUT,PATCH,DELETE

Like any other collection, relations support pagination, sorting and filtering.

It is also possible to have relations between the same type of resources, for example a relation "invitation" between "user"
and "user". The generated routes are as follows:

	/invitations GET,PUT,PATCH,DELETE with optional query parameters ?left={inviting_user_id} and ?right={invited_user_id}
	/invitations/{inviting_user_id}:{invited_users_id} GET,PUT,PATCH,DELETE

You can also query all relations of a specific user regardless of wether they are the inviting user or the invited user:

	/invitations?either={user_id}

Relations between the same type of resource can also be non-directional, by specifying the "non_directional" to true
in the configuration. For example the relation "friendship" between "user" and "user":

	/friendships GET,PUT,PATCH,DELETE with optional query parameters ?either={user_id}
	/friendships/{user_id}:{other_user_id} GET,PUT,PATCH,DELETE

The order between "user_id" and "other_user_id" does not matter. In the reported non-directional friendship object, one
will be "left_user_id" and one will be "right_user_id", depending on the lexicographic order of the identifiers.